Initial commit
This is the initial commit of the new webapp. The webapp is meant to
present online learning materials in form of slides with detailed
explanations. Kind of slides and a script together. For learning both
can be displayed side by side. For presentation the explanations can be
switched off.

In this version the navigation bar already works but neither the
slideshow mode nor the overview mode show any content.