Major refactoring of the event mechanism
This is a major refactoring of the code base with the aim to simplify
it. Instead of the previously used jQuery custom events, which were
manually triggered and acted upon, a simple data-binding scheme is used.
Therefor this commit introduces the `ObverservableValue` class, which
automatically calls a function, updates a DOM element or a jQuery
collection when its value changes. This allowed to remove many custom
update methods.

Also the commit splits the old `Slides` into the new classes
`Presentation` and `SlideshowPlayer` to follow the single responsibility
rule.

Next step will be support for the HTML5 history API. The methods are
already there but yet unimplemented.