Refactor to be a library instead of a stand-alone html app
This commit reverts the previous introduction of a `plugins.js` file in
favour of a much better approach. Instead of having clients clone this
repository in order to create presentations, make them is it as a
library.

The idea is, that in order to create a new presentation one creates a
new npm module and adds lecture-slides.js as a dependency. Therefor
users can easily up- and downgrade versions of the library and don't
have to track and merge back its source code changes.

Thinking of it, this is pretty neat. :-)