First simple version of the overview page
Now the overview page really lists all slides. One can click on a slide
title in order to call it up. Also a preview picture of the current
slide is show. The original plan was to have many small thumbnails
arranged in a grid. But it turns out that an HTML element retains its
bounding box after scaling it down.

The internet says one should wrap the element then, get the scaled
width/high of the inner element with getClientBoundingRect() and apply
it to the wrapper element. But as it seems this doesn't work either,
because the thumbails will then get scaled down again ...

Too much for me now. So a simple preview of the current slide it is.