Structuring decks with chapters and captions
This commit adds a new feature to structure a slide deck with chapters
and captions. Chapters are simple title slides which separate the
chapters of a deck. They are defined with the data-chapter attribute:

  <section data-chapter>
    …
  </section>

This will change the way the Table of Contents is rendered and will make
all subsequent slides be contents of that chapter.

Additionally slides may have captions which are also shown on the Table
of Contents. This applies to chapter slides as well as to normal slides:

  <section>
    <slide-caption>
      This will be shown in the Table of Contents below the slide title
    </slide-caption>
  </section>

The purpose of this is to add explanatory texts to the chapter titles.
But it can also be used to enumerate the content of normal slides.
For example like this:

  <section data-chapter>
    <h1>Front Matter</h1>
    <slide-caption>
      Welcome to this course. Here you'll find all that's needed to get
      you started.
    </slide-caption>
    …
  </section>

  <section>
    <h1>Syllabus</h1>
    <slide-caption>
      <ol>
        <li>Course Content</li>
        <li>Time Sheet</li>
        <li>Contact and Office Hours</li>
      </ol>
    </slide-caption>
    …
  </section>