Don't use mouse for touch gestures
This fixes a very annoying bug for desktop users. By default the
HammerJS library allows to simulate touch gestures with the mouse. While
great for development it is quite annoying for end-users. Because double
clicks or text selections were recognized as touch gestures. This made
it impossible to select a single line of text and triggered a lot of
unwanted navigations.

This simple fix now checks the pointer type of the gesture event. If the
pointer type is "mouse" the event will be discarded.